Update + new widget : Slider

+++ src/Cursor.cpp
@@ -1,16 +1,18 @@
 #include <hgui/header/Cursor.h>
 
-hgui::kernel::Cursor::Cursor(cursors standardCursor) : m_cursorPTR(glfwCreateStandardCursor(static_cast<int>(standardCursor)))
-{}
+hgui::kernel::Cursor::Cursor(cursors standardCursor) :
+	m_cursorPTR(glfwCreateStandardCursor(static_cast<int>(standardCursor)))
+{
+}
 
 hgui::kernel::Cursor::Cursor(const std::shared_ptr<Image>& customCursor, const point& clickPosition) : m_cursorPTR(nullptr)
 {
 	const GLFWimage cursor
-			{
-				.width = customCursor->get_data().width,
-				.height = customCursor->get_data().height,
-				.pixels = customCursor->get_data().pixels
-			};
+	{
+		.width = static_cast<int>(customCursor->get_data().width),
+		.height = static_cast<int>(customCursor->get_data().height),
+		.pixels = customCursor->get_data().pixels
+	};
 	if (cursor.pixels)
 	{
 		m_cursorPTR = glfwCreateCursor(&cursor, static_cast<int>(clickPosition.x), static_cast<int>(clickPosition.y));
